+++ client/doublezerod/internal/routing/config.go
@@ -6,34 +6,58 @@ import (
 	"log/slog"
 	"net"
 	"os"
+	"sync"
 )
 
 type RouteConfig struct {
 	Exclude []string `json:"exclude"`
 }
 
-type ConfiguredRouteReaderWriter struct {
-	log     *slog.Logger
-	nlr     Netlinker
+type ConfiguredRoutes struct {
+	cfg     *RouteConfig
 	exclude map[string]struct{}
+	mu      sync.Mutex
 }
 
-func NewConfiguredRouteReaderWriter(log *slog.Logger, nlr Netlinker, path string) (*ConfiguredRouteReaderWriter, error) {
+func NewConfiguredRoutes(path string) (*ConfiguredRoutes, error) {
 	cfg, err := loadConfig(path)
 	if err != nil {
 		return nil, fmt.Errorf("error loading route config: %v", err)
 	}
-	log.Info("routes: loaded routes", "routes", len(cfg.Exclude))
-	return &ConfiguredRouteReaderWriter{
-		log:     log,
-		nlr:     nlr,
+	return &ConfiguredRoutes{
+		cfg:     cfg,
 		exclude: makeExcludeMap(cfg.Exclude),
 	}, nil
 }
 
+func (c *ConfiguredRoutes) GetExcluded() map[string]struct{} {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	return c.exclude
+}
+
+func (c *ConfiguredRoutes) IsExcluded(ip string) bool {
+	_, ok := c.GetExcluded()[ip]
+	return ok
+}
+
+type ConfiguredRouteReaderWriter struct {
+	log *slog.Logger
+	nlr Netlinker
+	cfg *ConfiguredRoutes
+}
+
+func NewConfiguredRouteReaderWriter(log *slog.Logger, nlr Netlinker, cfg *ConfiguredRoutes) (*ConfiguredRouteReaderWriter, error) {
+	log.Info("routes: loaded routes", "excluded", len(cfg.GetExcluded()))
+	return &ConfiguredRouteReaderWriter{
+		log: log,
+		nlr: nlr,
+		cfg: cfg,
+	}, nil
+}
+
 func (c *ConfiguredRouteReaderWriter) RouteAdd(r *Route) error {
-	_, ok := c.exclude[r.Dst.IP.String()]
-	if ok {
+	if c.cfg.IsExcluded(r.Dst.IP.String()) {
 		c.log.Info("routes: excluding configured route from route add", "route", r.String())
 		return nil
 	}
@@ -41,8 +65,7 @@ func (c *ConfiguredRouteReaderWriter) RouteAdd(r *Route) error {
 }
 
 func (c *ConfiguredRouteReaderWriter) RouteDelete(r *Route) error {
-	_, ok := c.exclude[r.Dst.IP.String()]
-	if ok {
+	if c.cfg.IsExcluded(r.Dst.IP.String()) {
 		c.log.Info("routes: excluding configured route from route delete", "route", r.String())
 		return nil
 	}